collaborator access modal now uses useActions
