+++ src/client/components/CollaboratorAccessModal/hooks/index.ts
@@ -1,0 +1,1 @@
+export { toggleOption } from './useActions'
